Added (untested) solver kernels, fixed planes
I was wrong. It's done the other way. Oh well, this should fix
  planes being all over the place when attempted to set their free
  parameter to something other than zero.

LinearProgrammingProblemDisplay::solve() is now a thing. Somewhat.
I don't quite like how it's returning/storing the solution. And the
  matrix filling stuff, could be done better. So that is a subject
  to change. In a later commit, I'm tired and did fix a thing so counts.
It should compute though. Oh and with this commit, it's now much harder
  to compile this on windows, since cddlib. That needs to be fixed.

Attempted to make camera and orthography work. Didn't go well.
  Disabled until better days.