Added (hacky) localization support
It's not dynamic, meaning we don't actively scan for .mo files.
It's quite janky, meaning we use global variables and what not.
It's done the other way around, for when the system doesn't recognize
  the language. (and language reloading is a bit messy.)
  also i broke the line pretty :c
It might not work on Windows.

It will be improved but I need a save point.
Few thonks about font selection are needed. I'll include DejaVu
  Sans Mono font for now, but we might just go with built-in ones.
  Surely there is a common 'sans' font between three systems.